Module Bio_io.Btab_queries

Parsing btab files. Treats query as the record unit rather than individual hits.

module Record : sig ... end

A record type for Btab homology search files

module In_channel : sig ... end

In_channel for Btab files where each query in the file is a single record.