Module Btab_queries.Record

A record type for Btab homology search files

Overview

Unlike the Btab module, the query sequence is the basis for the record.

All hits are grouped by the query sequence. For example, in a btab file with query-target pairs, q1-t1, q1-t2, q2-t2 then you will have two records. One for q1 and one for q2.

API

type t

A btab query record. I.e., query name and a list of all Btab.t hits.

Creating

val create : Base.string -> Btab.Record.t Base.list -> t

create query_name hit_list creates a t given the name of the query and a list of hits. You probably won't use this function directly.

Accessing

val query : t -> Base.string

query t returns the name of the query for this record (i.e., for the list of hits).

val hits : t -> Btab.Record.t Base.list

hits t returns the list of hits associated with this query sequence.