Module Bio_io.Private

Internally used modules

module In_channel : sig ... end

Alternate interface to Stdio.In_channel. The create and with_file functions don't have a binary option, so it is compatible with the Record_in_channel.Make functor.

module Peekable_in_channel : sig ... end

A wrapper of Jane Street's Stdio.In_channel. Add's peek_char and peek_line functions that work on fifos as well as regular files.